+++ src/__figma__/logo-blau.mobile.figma.ts
@@ -1,0 +1,13 @@
+// url=https://www.figma.com/design/WCkDDzlXE16R6yXaljxddj/M%C3%ADstica-Mobile?node-id=43729-103
+// source=src/logo-blau.tsx
+// component=LogoBlau
+import figma from 'figma';
+
+export default {
+    example: figma.code`
+        <LogoBlau />
+    `,
+    imports: ['import {LogoBlau} from "@telefonica/mistica";'],
+    id: 'logo-blau-mobile',
+    metadata: {nestable: false},
+};